refactor: 불필요한 추상화 제거  - App 개행
