+++ src/main/java/com/devcourse/java/App.java
@@ -15,8 +15,7 @@ import com.devcourse.java.domain.menu.Menu;
 import com.devcourse.java.domain.menu.Menus;
 import com.devcourse.java.domain.menu.Query;
 import com.devcourse.java.domain.operator.Operator;
-import com.devcourse.java.domain.parser.ExpressionParser;
-import com.devcourse.java.domain.parser.PrefixParser;
+import com.devcourse.java.domain.calculator.parser.PrefixParser;
 import com.devcourse.java.domain.runner.CalculatorRunner;
 import com.devcourse.java.domain.storage.CalculateResult;
 import com.devcourse.java.domain.storage.MemoryStorage;
@@ -25,14 +24,17 @@ import com.devcourse.java.domain.storage.Storage;
 public class App {
     public static void main(String[] args) {
         Storage<CalculateResult> resultStorage = new MemoryStorage();
+
+        PrefixParser prefixParser = new PrefixParser();
+        Factory<Operator, String> operatorFactory = new OperatorFactory();
+        Calculator prefixCalculator = new PrefixCalculator(prefixParser, operatorFactory);
+
+        Query query = new Query(resultStorage);
+        Calculate calculate = new Calculate(prefixCalculator, resultStorage);
+
         Input reader = new Reader();
         Output writer = new Writer();
         Console console = new Console(reader, writer);
-        Query query = new Query(resultStorage);
-        ExpressionParser prefixParser = new PrefixParser();
-        Factory<Operator, String> operatorFactory = new OperatorFactory();
-        Calculator prefixCalculator = new PrefixCalculator(prefixParser, operatorFactory);
-        Calculate calculate = new Calculate(prefixCalculator, resultStorage);
         Factory<Menu, Menus> menuFactory = new MenuFactory(query, calculate);
 
         CalculatorRunner calculatorRunner = new CalculatorRunner(menuFactory, console);
