--- src/main/java/com/devcourse/java/domain/parser/ExpressionParser.java
@@ -1,7 +1,0 @@
-package com.devcourse.java.domain.parser;
-
-import java.util.List;
-
-public interface ExpressionParser {
-    List<String> parse(String expression);
-}
