+++ src/test/java/com/devcourse/java/domain/calculator/CalculatorTest.java
@@ -3,7 +3,7 @@ package com.devcourse.java.domain.calculator;
 import com.devcourse.java.domain.factory.Factory;
 import com.devcourse.java.domain.factory.OperatorFactory;
 import com.devcourse.java.domain.operator.Operator;
-import com.devcourse.java.domain.parser.PrefixParser;
+import com.devcourse.java.domain.calculator.parser.PrefixParser;
 import com.devcourse.java.domain.storage.CalculateResult;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
