+++ src/test/java/com/devcourse/java/domain/parser/ParserTest.java
@@ -1,5 +1,6 @@
 package com.devcourse.java.domain.parser;
 
+import com.devcourse.java.domain.calculator.parser.PrefixParser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
